Create dfa_mult_6.cpp
An implementation of a DFA using a 2d array containing the states and transitions. It verifies if a given binary string {0,1} is multiple of 6.
It prints out the transition table with the transitions between states containing the current character read to make it more clear to understand how a DFA works.